+++ changelog.md
@@ -4,6 +4,9 @@
 - Fixed calling the `onServerShutdown` event
 - Implemented loading plugins in the order of the queue
 - Added unique events (the only ones) with the return of the object: annotation `SubscribeSingleEvent` and method `invokeSingleEventAndReturn`
+- Added new custom events
+- Added API tools for player management and manipulation for the server
+- Added API tools for server management
 
 # FluxLoader v0.2.1
 - The system for subscribing classes as listeners has been fixed. Now it is not the class itself that is signed, but its instance.